requirements.txt
asgiref==3.8.1
certifi==2025.1.31
distlib==0.3.9
Django==5.2
djangorestframework==3.16.0
filelock==3.17.0
packaging==24.2
pillow==11.1.0
pipenv==2024.4.1
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.29.1
